% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_funs.R
\name{tract_sample}
\alias{tract_sample}
\title{Tract Sample}
\usage{
tract_sample(geoid, n)
}
\arguments{
\item{geoid}{Tract GEOID, n Sample size
#' @param n Sample size}
}
\description{
This function gets a sample of Sex-Age-Race.
}
\examples{
tract_sample()
}
